[refactor] generator 패키지 위치 변경 및 기능 수정
*generator 모든 generator가 domain과의 의존성을 없애기 위해 Integer List를 반환하도록 조정
