--- src/main/java/domain/lotto/generator/NumberGenerator.java
@@ -1,9 +1,0 @@
-package domain.lotto.generator;
-
-import domain.lotto.LottoNumber;
-
-import java.util.SortedSet;
-
-public interface NumberGenerator {
-    SortedSet<LottoNumber> create();
-}
