+++ src/main/java/generator/UserNumberGenerator.java
@@ -1,0 +1,12 @@
+package generator;
+
+import view.InputView;
+
+import java.util.List;
+
+public class UserNumberGenerator implements NumberGenerator {
+    @Override
+    public List<Integer> create() {
+        return InputView.inputNumbers();
+    }
+}
