+++ src/test/java/domain/lotto/LottoGameTest.java
@@ -5,6 +5,7 @@ import domain.lotto.lottoresult.LottoRank;
 import domain.lotto.lottoresult.LottoResult;
 import domain.lotto.lottoresult.LottoWinner;
 import domain.lotto.lottoresult.ResultCount;
+import generator.TestNumberGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
