+++ src/test/java/generator/TestNumberGenerator.java
@@ -1,34 +1,28 @@
-package domain.lotto;
-
-import domain.lotto.generator.NumberGenerator;
+package generator;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.SortedSet;
-import java.util.TreeSet;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class TestNumberGenerator implements NumberGenerator {
-    private final SortedSet<LottoNumber> numbers;
+    private final List<Integer> numbers;
 
     public TestNumberGenerator() {
         numbers = IntStream.range(1, 7)
-                .mapToObj(LottoNumberFactory::getInstance)
-                .collect(Collectors.toCollection(TreeSet::new));
+                .mapToObj(Integer::new)
+                .collect(Collectors.toList());
     }
 
     public TestNumberGenerator(List<Integer> values) {
         if (Objects.isNull(values)) {
             throw new IllegalArgumentException("null값은 입력할 수 없습니다.");
         }
-        this.numbers = values.stream()
-                .map(LottoNumberFactory::getInstance)
-                .collect(Collectors.toCollection(TreeSet::new));
+        this.numbers = values;
     }
 
     @Override
-    public SortedSet<LottoNumber> create() {
+    public List<Integer> create() {
         return numbers;
     }
 }
